Moved SkillType to entity package
