+++ src/main/java/com/scoreit/entity/SkillType.java
@@ -1,4 +1,4 @@
-package com.scoreit;
+package com.scoreit.entity;
 
 public enum SkillType {
     BATSMAN,
